+++ http/src/main/scala/com/twitter/finatra/http/filters/HttpResponseFilter.scala
@@ -1,13 +1,14 @@
 package com.twitter.finatra.http.filters
 
 import com.twitter.conversions.DurationOps._
-import com.twitter.finagle.http.{Fields, MediaType, Message, Request, Response}
+import com.twitter.finagle.http.{Fields, HeaderMap, MediaType, Message, Request, Response}
 import com.twitter.finagle.{Service, SimpleFilter}
 import com.twitter.finatra.http.request.RequestUtils
 import com.twitter.inject.Logging
 import com.twitter.util.{Future, Return, ScheduledThreadPoolTimer, Throw, Try}
 import java.net.URI
 import javax.inject.Singleton
+import scala.util.control.NonFatal
 
 /**
  * HttpResponseFilter does the following:
@@ -18,8 +19,15 @@ import javax.inject.Singleton
  * order to always attempt to fully specify a relative reference, this class should be instantiated
  * with the constructor arg [[fullyQualifyLocationHeader]] set to 'true'.
  *
- *  @param fullyQualifyLocationHeader the Filter should attempt to always fully qualify
+ * @param fullyQualifyLocationHeader the Filter should attempt to always fully qualify
  *                                    the value of the response 'Location' header. Default: false
+ *
+ * @note This filter does NOT throw exceptions when it is unable to set a location
+ *       header value because of non RFC 7230 compliant values. Generally this filter is
+ *       installed "above" the `ExceptionMappingFilter` and therefore translation of an exception
+ *       into an appropriate HTTP response is not available thus the filter logs a warning message.
+ *       This means that if there is a non-compliant 'Location' header value in the outgoing
+ *       Response this filter will not fully qualify, nor replace, nor remove it.
  *
  * @see [[https://en.wikipedia.org/wiki/HTTP_location HTTP location]]
  * @see [[https://tools.ietf.org/html/rfc7231 Hypertext Transfer Protocol (HTTP/1.1): Semantics and Content]]
@@ -54,6 +62,26 @@ class HttpResponseFilter[R <: Request](fullyQualifyLocationHeader: Boolean)
     }
   }
 
+  /* Protected */
+
+  /**
+   * Allows extensions to set a different function for setting the Response Location Header
+   * value, e.g., `setUnsafe` (which performs no validation) instead of the default `set`
+   * (which performs validation).
+   * @param headerMap the [[com.twitter.finagle.http.HeaderMap]] on which to set the Location header
+   * @param value the Location header value to set.
+   * @return the set function to use for setting the value.
+   */
+  protected def setLocationHeader(headerMap: HeaderMap, value: String): Unit = {
+    try {
+      headerMap.set(Fields.Location, value)
+    } catch {
+      case NonFatal(e) =>
+        warn(
+          s"Unable to set Response Header '${Fields.Location}' value: '$value' - ${e.getMessage}")
+    }
+  }
+
   /* Private */
 
   /**
@@ -66,7 +94,7 @@ class HttpResponseFilter[R <: Request](fullyQualifyLocationHeader: Boolean)
    * @see Server: [[https://tools.ietf.org/html/rfc7231#section-7.4.2 Section 7.4.2 of RFC 7231]]
    * @see Content-Type: [[https://tools.ietf.org/html/rfc7231#section-3.1.1.5 Section 3.1.1.5 of RFC 7231]]
    */
-  private def setResponseHeaders(response: Response): Unit = {
+  private[this] def setResponseHeaders(response: Response): Unit = {
     response.headerMap.setUnsafe(Fields.Server, "Finatra")
     response.headerMap.setUnsafe(Fields.Date, currentDateValue)
     if (response.contentType.isEmpty && response.length != 0) {
@@ -75,28 +103,26 @@ class HttpResponseFilter[R <: Request](fullyQualifyLocationHeader: Boolean)
     }
   }
 
-  private def getCurrentDateValue: String = Message.httpDateFormat(System.currentTimeMillis())
+  private[this] def getCurrentDateValue: String = Message.httpDateFormat(System.currentTimeMillis())
 
-  private def updateLocationHeader(request: R, response: Response): Unit = {
+  private[this] def updateLocationHeader(request: R, response: Response): Unit = {
     for (existingLocation <- response.location) {
       Try(new URI(existingLocation)) match {
         case Throw(e) =>
           warn(
-            s"Response location header value $existingLocation is not a valid URI. ${e.getMessage}"
+            s"Response Header '${Fields.Location}' value: '$existingLocation' is not a valid URI. ${e.getMessage}"
           )
         case Return(uri) if isRelativeLocation(uri) =>
-          response.headerMap.set(
-            Fields.Location,
-            getLocationHeaderValue(fullyQualify = this.fullyQualifyLocationHeader, uri, request)
-          )
+          setLocationHeader(
+            response.headerMap,
+            getLocationHeaderValue(fullyQualify = this.fullyQualifyLocationHeader, uri, request))
         case Return(uri) =>
           // valid URI with a non-null scheme, check scheme against "x-forwarded-proto" from request
           request.headerMap.get("x-forwarded-proto") match {
             case Some(protocol) if protocol != uri.getScheme =>
-              response.headerMap.set(
-                Fields.Location,
-                getLocationHeaderValue(fullyQualify = true, uri, request)
-              )
+              setLocationHeader(
+                response.headerMap,
+                getLocationHeaderValue(fullyQualify = true, uri, request))
             case _ =>
             // valid URI with a non-null scheme, no "x-forwarded-proto" from request = do nothing
           }
@@ -107,7 +133,11 @@ class HttpResponseFilter[R <: Request](fullyQualifyLocationHeader: Boolean)
   // uri.getScheme == null is a relative reference
   private[this] def isRelativeLocation(uri: URI): Boolean = uri.getScheme == null
 
-  private def getLocationHeaderValue(fullyQualify: Boolean, uri: URI, request: Request): String = {
+  private[this] def getLocationHeaderValue(
+    fullyQualify: Boolean,
+    uri: URI,
+    request: Request
+  ): String = {
     val normalizedURI = uri.normalize()
 
     val scheme: String =